Make SigningController an interface
